+++ RedPandaIDE/settingsdialog/editorgeneralwidget.cpp
@@ -40,7 +40,7 @@ void EditorGeneralWidget::doLoad()
     //indents
     ui->chkAddIndent->setChecked(pSettings->editor().addIndent());
     ui->chkAutoIndent->setChecked(pSettings->editor().autoIndent());
-    ui->chkTabToSpaces->setChecked(pSettings->editor().autoIndent());
+    ui->chkTabToSpaces->setChecked(pSettings->editor().tabToSpaces());
     ui->spTabWidth->setValue(pSettings->editor().tabWidth());
     ui->chkShowIndentLines->setChecked(pSettings->editor().showIndentLines());
     ui->colorIndentLine->setColor(pSettings->editor().indentLineColor());
@@ -72,7 +72,7 @@ void EditorGeneralWidget::doSave()
     //indents
     pSettings->editor().setAddIndent(ui->chkAddIndent->isChecked());
     pSettings->editor().setAutoIndent(ui->chkAutoIndent->isChecked());
-    pSettings->editor().setAutoIndent(ui->chkTabToSpaces->isChecked());
+    pSettings->editor().setTabToSpaces(ui->chkTabToSpaces->isChecked());
     pSettings->editor().setTabWidth(ui->spTabWidth->value());
     pSettings->editor().setShowIndentLines(ui->chkShowIndentLines->isChecked());
     pSettings->editor().setIndentLineColor(ui->colorIndentLine->color());